backend/requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
djoser==2.2.0
drf-yasg==1.21.5
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
